+++ server/src/macroLanguageService/parser/macroNodes.ts
@@ -4,7 +4,6 @@
  *--------------------------------------------------------------------------------------------*/
 'use strict';
 
-import { basename } from 'path';
 import { integer } from 'vscode-languageserver';
 
 export enum ReferenceType {
@@ -25,9 +24,9 @@ export enum ValueAttribute {
 	Signed,
 	GCode,
 	MCode,
-	Parameter
+	Parameter,
+	Program
 }
-
 
 export function getNodeAtOffset(node: Node, offset: number): Node | null {
 
@@ -44,7 +43,7 @@ export function getNodeAtOffset(node: Node, offset: number): Node | null {
 		if (node.offset <= offset && node.end >= offset) {
 			if (!candidate) {
 				candidate = node;
-			} else if (node.length <= candidate.length || node.offset >= candidate.offset+candidate.length) {
+			} else if (node.length <= candidate.length || node.offset >= candidate.offset + candidate.length) {
 				candidate = node;
 			}
 			return true;
@@ -71,41 +70,37 @@ export interface ITextProvider {
 	(offset: number, length: number): string;
 }
 
-export interface ISymbolLink {
-	symNode: Symbol | Label;
-	defType: NodeType;
-	valNode: Node;	
-}
-
 export class Node {
 
 	public parent: Node | null;
 
 	public offset: number;
 	public length: number;
-	private _symbolLink: ISymbolLink | undefined;
+	public progOffset: number;
+	public progLength: number;
+
+	public symbol: Symbol | Label;
 
 	public get end() { return this.offset + this.length; }
 
 	public options: { [name: string]: any; } | undefined;
 
 	public textProvider: ITextProvider | undefined; // only set on the root node
+	public textProviderProg: ITextProvider | undefined; // only set on the root node
 
 	private children: Node[] | undefined;
 	private issues: IMarker[] | undefined;
 
 	private nodeType: NodeType | undefined;
 
-	constructor(offset: number = -1, len: number = -1, symbol?: ISymbolLink, nodeType?: NodeType) {
+	constructor(offset: number = -1, len: number = -1, progOffset: number = -1, progLen: number = -1, nodeType?: NodeType) {
 		this.parent = null;
 		this.offset = offset;
 		this.length = len;
+		this.progOffset = progOffset;
+		this.progLength = progLen;
 		if (nodeType) {
 			this.nodeType = nodeType;
-		}
-
-		if (symbol) {
-			this._symbolLink = symbol;
 		}
 	}
 
@@ -116,16 +111,8 @@ export class Node {
 	public get type(): NodeType {
 		return this.nodeType || NodeType.Undefined;
 	}
-	
-	public get symbolLink(): ISymbolLink {
-    	return this._symbolLink;
-	}
 
-	public set symbolLink(symbol: ISymbolLink) {
-		this._symbolLink = symbol;
-	}
-
-	protected getTextProvider(): ITextProvider {
+	public getTextProvider(): ITextProvider {
 		let node: Node | null = this;
 		while (node && !node.textProvider) {
 			node = node.parent;
@@ -136,12 +123,23 @@ export class Node {
 		return () => { return 'unknown'; };
 	}
 
-	public getText(): string {
-		return this.symbolLink?.valNode.getText() ?? this.getTextProvider()(this.offset, this.length);
+	public getTextProviderProg(): ITextProvider {
+		let node: Node | null = this;
+		while (node && !node.textProviderProg) {
+			node = node.parent;
+		}
+		if (node) {
+			return node.textProviderProg!;
+		}
+		return () => { return 'unknown'; };
 	}
 
-	public getNodeText(): string {
+	public getText(): string {
 		return this.getTextProvider()(this.offset, this.length);
+	}
+
+	public getNonSymbolText(): string {
+		return this.getTextProviderProg()(this.progOffset, this.progLength);
 	}
 
 	public matches(str: string): boolean {
@@ -332,7 +330,7 @@ export class Node {
 		if (this.parent.hasChildren()) {
 			const index = this.getChildIndex();
 			if (index !== null) {
-				return this.parent.getChild(index-1);
+				return this.parent.getChild(index - 1);
 			}
 		}
 		return null;
@@ -374,8 +372,8 @@ export class Node {
 }
 
 export interface NodeConstructor<T> {
-	new(offset: number, len: number, symbol?: ISymbolLink): T;
-	
+	new(offset: number, len: number, progOffset: number, progLen: number): T;
+
 }
 
 export interface IRule {
@@ -474,11 +472,11 @@ export class ParseErrorCollector implements IVisitor {
 	}
 }
 
-
 export enum NodeType {
 	Undefined,
 	MacroFile,
 	DefFile,
+	SymbolRoot,
 	Include,
 	StringLiteral,
 	Program,
@@ -513,27 +511,54 @@ export enum NodeType {
 	SequenceNumber,
 	NNAddress,
 	BlockSkip,
-	BlockDel
+	BlockDel,
 }
 
-export class Reference extends Node { 
+export class Reference extends Node {
 
 	public referenceTypes: ReferenceType;
+
+	constructor(offset: number, length: number, progOffset: number, ProgLength: number) {
+		super(offset, length, progOffset, ProgLength);
+	}
 
 	public addReferenceType(...referenceTypes: ReferenceType[]) {
 		for (const ref of referenceTypes) {
 			this.referenceTypes |= ref;
-			if (this.symbolLink) {
-				this.symbolLink.symNode.referenceTypes |= ref;
+			if (this.symbol) {
+				this.symbol.referenceTypes |= ref;
 			}
 		}
 	}
 
-	public hasReferenceType(referenceType:ReferenceType): boolean {
+	public hasReferenceType(referenceType: ReferenceType): boolean {
 		return !!(this.referenceTypes & referenceType);
 	}
 }
 
+export class DefReference extends Reference {
+
+	public attrib: ValueAttribute = ValueAttribute.None;
+
+	constructor(offset: number, length: number, progOffset: number, ProgLength: number, private definition: AbstractDefinition) {
+		super(offset, length, progOffset, ProgLength);
+		if (definition) {
+			this.attrib = definition.attrib;
+		}
+	}
+
+	public get valueType(): NodeType | undefined {
+		return this.definition?.value?.type;
+	}
+
+	public get defType(): NodeType | undefined {
+		return this.definition?.type;
+	}
+
+	public getNonSymbolText(): string {
+		return this.definition?.value?.getText();
+	}
+}
 
 export class MacroFile extends Node {
 
@@ -564,8 +589,8 @@ export class Include extends Node {
 
 export class NcStatement extends Node {
 
-	constructor(offset: number, length: number, symbol: ISymbolLink) {
-		super(offset, length, symbol);
+	constructor(offset: number, length: number) {
+		super(offset, length);
 	}
 
 	public get type(): NodeType {
@@ -580,11 +605,11 @@ export enum CodeType {
 
 export class NcCode extends Reference {
 
-	public referenceTypes: ReferenceType = ReferenceType.Code | ReferenceType.Symbol;
-	public codeType?:CodeType;
+	public codeType?: CodeType;
 
-	constructor(offset: number, length: number) {
-		super(offset, length);
+	constructor(offset: number, length: number, progOffset: number, progLength: number) {
+		super(offset, length, progOffset, progLength);
+		this.referenceTypes = ReferenceType.Code;
 	}
 
 	public get type(): NodeType {
@@ -594,10 +619,9 @@ export class NcCode extends Reference {
 
 export class Parameter extends Reference {
 
-	public referenceTypes: ReferenceType = ReferenceType.Symbol;
-
-	constructor(offset: number, length: number) {
-		super(offset, length);
+	constructor(offset: number, length: number, progOffset: number, ProgLength: number) {
+		super(offset, length, progOffset, ProgLength);
+		this.referenceTypes = ReferenceType.Symbol;
 	}
 
 	public get type(): NodeType {
@@ -607,11 +631,11 @@ export class Parameter extends Reference {
 
 export class SequenceNumber extends Reference {
 
-	public referenceTypes: ReferenceType = ReferenceType.Symbol;
 	public number?: Node;
 
-	constructor(offset: number, length: number) {
-		super(offset, length);
+	constructor(offset: number, length: number, progOffset: number, progLength: number) {
+		super(offset, length, progOffset, progLength);
+		this.referenceTypes = ReferenceType.Sequence;
 	}
 
 	public get type(): NodeType {
@@ -629,16 +653,16 @@ export class SequenceNumber extends Reference {
 
 export class BodyDeclaration extends Node {
 
-	constructor(offset: number, length: number, symbol?: ISymbolLink) {
-		super(offset, length, symbol);
+	constructor(offset: number, length: number) {
+		super(offset, length);
 	}
 }
 
 export class Program extends BodyDeclaration {
 	public identifier?: Node;
 
-	constructor(offset: number, length: number, symbol:ISymbolLink) {
-		super(offset, length, symbol);
+	constructor(offset: number, length: number) {
+		super(offset, length);
 	}
 
 	public get type(): NodeType {
@@ -658,21 +682,20 @@ export class Program extends BodyDeclaration {
 	}
 }
 
-export class Symbol extends Reference {
-	public referenceTypes: ReferenceType = ReferenceType.Symbol;
-	public valueType: NodeType;
-	public attrib: ValueAttribute = ValueAttribute.None;
+export class Symbol extends DefReference {
+
 	public identifier: Node;
 
-	constructor(offset: number, length: number) {
-		super(offset, length);
+	constructor(offset: number, length: number, progOffset: number, ProgLength: number, definition?: AbstractDefinition) {
+		super(offset, length, progOffset, ProgLength, definition);
 		this.setIdentifier(new Node(offset, length));
+		this.referenceTypes = ReferenceType.Symbol;
 	}
 
 	public get type(): NodeType {
 		return NodeType.Symbol;
 	}
-	
+
 	public setIdentifier(node: Node | null): node is Node {
 		return this.setNode('identifier', node, 0);
 	}
@@ -682,15 +705,14 @@ export class Symbol extends Reference {
 	}
 }
 
-export class Label extends Reference {
-	public referenceTypes: ReferenceType = ReferenceType.Label;
-	public valueType: NodeType;
-	public attrib: ValueAttribute = ValueAttribute.None;
+export class Label extends DefReference {
+
 	public identifier: Node;
 
-	constructor(offset: number, length: number) {
-		super(offset, length);
+	constructor(offset: number, length: number, progOffset: number, ProgLength: number, definition?: AbstractDefinition) {
+		super(offset, length, progOffset, ProgLength, definition);
 		this.setIdentifier(new Node(offset, length));
+		this.referenceTypes = ReferenceType.Label;
 	}
 
 	public get type(): NodeType {
@@ -704,18 +726,12 @@ export class Label extends Reference {
 	public getText(): string {
 		return this.identifier.getText();
 	}
-
-	public getNodeText(): string {
-		return super.getText();
-	}
 }
 
-export class Numeric extends Reference  {
+export class Numeric extends Reference {
 
-	public referenceTypes: ReferenceType;
-
-	constructor(offset: number, length: number, symbol: ISymbolLink) {
-		super(offset, length, symbol);
+	constructor(offset: number, length: number, progOffset: number, ProgLength: number) {
+		super(offset, length, progOffset, ProgLength);
 	}
 
 	public get type(): NodeType {
@@ -725,11 +741,11 @@ export class Numeric extends Reference  {
 
 export class Variable extends Reference {
 
-	public referenceTypes: ReferenceType = ReferenceType.Variable;
 	public body?: Node;
 
-	constructor(offset: number, length: number, symbol: ISymbolLink) {
-		super(offset, length, symbol);
+	constructor(offset: number, length: number, progOffset: number, ProgLength: number) {
+		super(offset, length, progOffset, ProgLength);
+		this.referenceTypes = ReferenceType.Variable;
 	}
 
 	public get type(): NodeType {
@@ -751,10 +767,9 @@ export class Variable extends Reference {
 
 export class Address extends Reference {
 
-	public referenceTypes: ReferenceType = ReferenceType.Address | ReferenceType.Symbol;
-
-	constructor(offset: number, length: number, symbol:ISymbolLink) {
-		super(offset, length, symbol);
+	constructor(offset: number, length: number, progOffset: number, progLength: number) {
+		super(offset, length, progOffset, progLength);
+		this.referenceTypes = ReferenceType.Address;
 	}
 
 	public get type(): NodeType {
@@ -766,8 +781,8 @@ export class Ffunc extends Node {
 
 	public identifier?: Node;
 
-	constructor(offset: number, length: number, symbol:ISymbolLink) {
-		super(offset, length, symbol);
+	constructor(offset: number, length: number) {
+		super(offset, length);
 	}
 
 	public get type(): NodeType {
@@ -785,8 +800,8 @@ export class Ffunc extends Node {
 
 export class Fcmd extends Ffunc {
 
-	constructor(offset: number, length: number, symbol:ISymbolLink) {
-		super(offset, length, symbol);
+	constructor(offset: number, length: number) {
+		super(offset, length);
 	}
 
 	public get type(): NodeType {
@@ -796,7 +811,7 @@ export class Fcmd extends Ffunc {
 
 export class BlockDel extends Node {
 
-	public number?:Numeric;
+	public number?: Numeric;
 
 	constructor(offset: number, length: number) {
 		super(offset, length);
@@ -813,14 +828,14 @@ export class BlockDel extends Node {
 	public getNumber(): Numeric | undefined {
 		return this.number;
 	}
-} 
+}
 
 export class GotoStatement extends BodyDeclaration {
 
-	public label?:Node;
-	
-	constructor(offset: number, length: number, symbol:ISymbolLink) {
-		super(offset, length, symbol);
+	public label?: Node;
+
+	constructor(offset: number, length: number) {
+		super(offset, length);
 	}
 
 	public get type(): NodeType {
@@ -887,8 +902,8 @@ export class ElseTermStatement extends ElseStatement {
 export class ConditionalStatement extends BodyDeclaration {
 	public contitional?: ConditionalExpression;
 
-	constructor(offset: number, length: number, symbol?:ISymbolLink) {
-		super(offset, length, symbol);
+	constructor(offset: number, length: number) {
+		super(offset, length);
 	}
 
 	public setConditional(node: ConditionalExpression | null): node is ConditionalExpression {
@@ -901,8 +916,8 @@ export class ConditionalStatement extends BodyDeclaration {
 }
 
 export class IfStatement extends ConditionalStatement {
-	constructor(offset: number, length: number, symbol:ISymbolLink) {
-		super(offset, length, symbol);
+	constructor(offset: number, length: number) {
+		super(offset, length);
 	}
 
 	public get type(): NodeType {
@@ -911,8 +926,8 @@ export class IfStatement extends ConditionalStatement {
 }
 
 export class WhileStatement extends ConditionalStatement {
-	public dolabel?:Node;
-	public endlabel?:Node;
+	public dolabel?: Node;
+	public endlabel?: Node;
 
 	constructor(offset: number, length: number) {
 		super(offset, length);
@@ -939,8 +954,8 @@ export class ConditionalExpression extends Node {
 	public condition?: Node;
 	public logic?: Node;
 
-	constructor(offset: number, length: number, symbol:ISymbolLink) {
-		super(offset, length, symbol);
+	constructor(offset: number, length: number) {
+		super(offset, length);
 	}
 
 	public get type(): NodeType {
@@ -994,8 +1009,8 @@ export class BinaryExpression extends Node {
 	public right?: Node;
 	public operator?: Node;
 
-	constructor(offset: number, length: number, symbol:ISymbolLink) {
-		super(offset, length, symbol);
+	constructor(offset: number, length: number) {
+		super(offset, length);
 	}
 
 	public get type(): NodeType {
@@ -1062,8 +1077,8 @@ export class Assignment extends Node {
 	public left?: Node;
 	public right?: Node;
 
-	constructor(offset: number, length: number, symbol:ISymbolLink) {
-		super(offset, length, symbol);
+	constructor(offset: number, length: number) {
+		super(offset, length);
 	}
 
 	public get type(): NodeType {
@@ -1108,7 +1123,7 @@ export class AbstractDefinition extends Node {
 	public getName(): string {
 		return this.identifier ? this.identifier.getText() : '';
 	}
-	
+
 	public setValue(node: Node | null): node is Node {
 		return this.setNode('value', node, 0);
 	}