+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.3.4  (-)
 - Function call hierarchy
+- Prepare rename provider for better rename experience
 
 ## 0.3.3  (April 16, 2021)
 - NN address parsing